requirements-dev.txt
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
anyio==4.2.0
aiosqlite==0.19.0
ruff==0.9.9
fakeredis[aioredis]==2.21.1
pyarrow==23.0.1
pandas==3.0.1
